% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_glossary}
\alias{get_glossary}
\title{Get all of the terms and corresponding information in the current glossary}
\usage{
get_glossary(select_terms = NULL, data_frame = FALSE)
}
\arguments{
\item{select_terms}{A character vector with terms corresponding to those in the glossary. If this is left NULL, all terms in the glossary will be retrieved.}
}
\value{
A list where each element is a list for each term, which is comrised of nested lists with relevant information.
}
\description{
Get all of the terms and corresponding information in the current glossary
}
