% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{list_categories}
\alias{list_categories}
\title{List the categories in the current glossary}
\usage{
list_categories()
}
\value{
A list where each element is a list for each term, which is comrised of nested lists with relevant information.
}
\description{
List the categories in the current glossary
}
