% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{list_terms}
\alias{list_terms}
\title{List the terms in the current glossary}
\usage{
list_terms(select_categories = NULL)
}
\arguments{
\item{categories}{To only retrieve the terms from particular categories, a character vector of those categories can be given. Otherwise, all terms from all categories will be returned as separate lists for each category. To see the possible categories, use the list_categories() function.}
}
\value{
A list where each element is a list for each category of terms
}
\description{
List the terms in the current glossary
}
